% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correally-package.R
\docType{package}
\name{correally-package}
\alias{correally}
\alias{correally-package}
\title{correally: Visualizing correlations}
\description{
Provides methods to create interactive visualizations of correlations based on plotly.
}
\details{
Description of what the package does in general
}
\section{correallyPlot}{


This is the main function of the package. It provides a wrapper to plotly to create an interactive visualization of a correlation matrix.
It provides a clean implementation for plotting correlation matrices, easy use of different color schemes
(Base R colors, Hex colors, \link{RColorBrewer} palettes as well as support for \link{viridisLite}). It also provides coloring options for backgrounds, grids and fonts
matching different \link{shinythemes}.
}

\section{correallyPlotAnimated}{

Not yet implemented
}

\section{Helper functions}{

\link{correally} comes with a bunch of helper functions that facilitate the main \link{correallyPlot} function.

\enumerate{
\item Validation functions
\itemize{
\item \link{validateBaseColor}
\item \link{validateHex}
\item \link{validateRColorBrewer}
\item \link{validateViridis}
}
}
}

\author{
\strong{Maintainer}: Stefan Haring \email{stefan.haring@yahoo.com}

}
\keyword{internal}
