% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correally.R
\name{validateBaseColor}
\alias{validateBaseColor}
\title{Validate an R Base color representation.}
\usage{
validateBaseColor(string)
}
\arguments{
\item{string}{String vector.}
}
\value{
Returns a logical vector of the same length as the \code{string} input.
}
\description{
\code{validateBaseColor} returns if a string vector contains color representations contained in the R Base colors.
}
\details{
This function validates a given R Base color representation.
It validates the input against the \code{\link[grDevices:colors]{grDevices::colors()}} function.
}
\examples{
validateBaseColor("red")
validateBaseColor("RED")

validateBaseColor(c("red", "blue"))

}
