% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correally.R
\name{validateHex}
\alias{validateHex}
\title{Validate a hexadecimal color representation.}
\usage{
validateHex(string, shorthand = FALSE, ignore.case = FALSE)
}
\arguments{
\item{string}{String vector.}

\item{shorthand}{A logical scalar. If \emph{TRUE}, the validation is performed against the shorthand notation and if \emph{FALSE},
validation is performed against the normal notation.}

\item{ignore.case}{A logical scalar. If \emph{FALSE}, the validation is case sensitive and if \emph{TRUE}, case is ignored during validation.}
}
\value{
Returns a logical vector of the same length as the \code{string} input.
}
\description{
\code{validateHex} returns if a string vector contains color representations in hexadecimal format.
}
\details{
This function validates a given hexadecimal color representation.
It can check the normal hexadecimal notation (e.g. #FFFFFF) or the shorthand notation (e.g. #FFF).
}
\examples{
validateHex("#FFFFFF")
validateHex("#FFF")
validateHex("#FFF", shorthand = TRUE)

validateHex("#ffffff")
validateHex("#ffffff", ignore.case = TRUE)

validateHex(c("#FFFFFF","asdf"))

}
