% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correally.R
\name{validateRColorBrewer}
\alias{validateRColorBrewer}
\title{Validate an RColorBrewer representation.}
\usage{
validateRColorBrewer(string)
}
\arguments{
\item{string}{String vector.}
}
\value{
Returns a logical vector of the same length as the \code{string} input.
}
\description{
\code{validateRColorBrewer} returns if a string vector contains color representations contained in the RColorBrewer package.
}
\details{
This function validates a given RColorBrewer representation.
It validates the input against the \link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info} method.
}
\examples{
validateRColorBrewer("RdBu")
validateRColorBrewer("RED")

validateRColorBrewer(c("BrBG", "blue"))

}
