% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correally.R
\name{validateViridis}
\alias{validateViridis}
\title{Validate a viridis color representation.}
\usage{
validateViridis(string)
}
\arguments{
\item{string}{String vector.}
}
\value{
Returns a logical vector of the same length as the \code{string} input.
}
\description{
\code{validateViridis} returns if a string vector contains color representations contained in the viridis package.
}
\details{
This function validates a given viridis representation. Valid inputs are \emph{viridis, magma, plasma, inferno} or \emph{cividis}.
They are part of the \link{viridis} package. More information can be found here: \link[viridisLite:viridis]{viridisLite::viridis}.
}
\examples{
validateViridis("viridis")
validateViridis("RED")

validateViridis(c("magma", "plasma"))

}
